app/tasks/pyproject.toml
[project]
name = "tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncmy>=0.2.11",
    "fastapi>=0.128.1",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "alembic>=1.18.3",
    "fastapi[standard]>=0.128.1",
]
